dateRange, priceRange y cityCountry validation and exceptions and some other minimal changes

+++ backend/src/main/java/com/queHacer/queHacer/Event/Model/EventDTO.java
@@ -10,8 +10,8 @@ public class EventDTO {
     private Integer id;
     private String name;
     private String description;
-    private Long minPrice;
-    private Long maxPrice;
+    private Float minPrice;
+    private Float maxPrice;
     private String instagram;
     private String facebook;
     private String whatsapp;
