+++ backend/src/main/java/com/queHacer/queHacer/Event/Model/UpdateEventDTO.java
@@ -9,8 +9,8 @@ import java.time.LocalDateTime;
 public class UpdateEventDTO {
     private String name;
     private String description;
-    private Long minPrice;
-    private Long maxPrice;
+    private Float minPrice;
+    private Float maxPrice;
     private String instagram;
     private String facebook;
     private String whatsapp;
